EE8084-Network Security
The attached is the final thesis of an assigned topic : Return-Oriented Programming to Compromise the iPhone. The word limits is 6000.

We met at least once a week to discuss the paper. I took charge of abstract, introduction,literature review and conclusion. 

Teammates : Hailin Chen, Luoyuan Xiong, Shen Yuan, Shengping Cui and Shuangchen Huang.